add cart

+++ addToCart.js
@@ -1,4 +1,8 @@
 import { getCartProductFromLS } from "./getCartProduct";
+import { updateCartValue } from "./updateCartValue";
+
+
+getCartProductFromLS();
 export const addToCart = (event, id, stock)=>{
 
     let arrLocalStorageProduct = getCartProductFromLS();
@@ -6,14 +10,37 @@ export const addToCart = (event, id, stock)=>{
     const currentProdElem = document.querySelector(`#card${id}`);
     let quantity = currentProdElem.querySelector(".productQuantity").innerText;
     let price = currentProdElem.querySelector(".productPrice").innerText;
-    console.log( quantity, price);
+    // console.log( quantity, price);
 
 
     price = price.replace("₹", "");
+
+    let existingProd = arrLocalStorageProduct.find((curProd) => curProd.id === id);
+    
+    // console.log(existingProd);
+  
+if(existingProd && quantity>1){
+    quantity= Number(existingProd.quantity) + Number(quantity);
+    price=Number(price * quantity);
+    let updatedCart ={id, quantity, price};
+
+    updatedCart = arrLocalStorageProduct.map((curProd) =>{
+        return curProd.id === id ? updatedCart : curProd;
+    });
+    localStorage.setItem("cartProductLS", JSON.stringify(updatedCart));
+}
+
+if (existingProd){
+    return false;
+}
+
      price = Number(price * quantity);
   quantity = Number(quantity);
 
   arrLocalStorageProduct.push({ id, quantity, price });
   localStorage.setItem("cartProductLS", JSON.stringify(arrLocalStorageProduct));
+
+  updateCartValue(arrLocalStorageProduct);
+  
   
 };
